universidad/requirements.txt
arabic-reshaper==3.0.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
binaryornot==0.4.4
boto3==1.28.3
botocore==1.31.3
Brotli==1.1.0
certifi==2023.5.7
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cookiecutter==2.1.1
crispy-tailwind==0.5.0
cryptography==43.0.1
cssselect2==0.7.0
dj-database-url==2.0.0
dj-static==0.0.6
Django==4.2.3
django-appconf==1.0.5
django-chartjs==2.3.0
django-ckeditor==6.6.1
django-crispy-forms==2.1
django-js-asset==2.1.0
django-model-utils==4.3.1
django-node==4.0.2
django-on-heroku==1.1.2
django-staticfiles==1.2.1
django-storages==1.13.2
django-tailwind==3.6.0
fonttools==4.54.1
gunicorn==20.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
lxml==5.3.0
MarkupSafe==2.1.3
oscrypto==1.3.0
Pillow==10.0.0
pipp==0.0.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.1
PyPDF2==3.0.1
pyphen==0.16.0
python-bidi==0.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
python-slugify==8.0.1
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.31.0
s3transfer==0.6.1
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
static3==0.7.0
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.3.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
Unipath==1.1
uritools==4.0.3
urllib3==1.26.16
weasyprint==62.3
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.4
zopfli==0.2.3
